Define CSS variables for page colors
Define CSS variables for all the colors used on the page in a ruleset for the `:root` pseudo-class, each with a comment indicating how and where the color is used.

The aim is to make the process of altering the color palette more efficient, with the user being able to do it from one ruleset rather than having to go through multiple ones throughout the stylesheet.